pio_usb: mask endp with b'111 (d'7)
This does not fix any issue, but makes a token packet creation
according to the usb 2.0 spec, where the third byte contains:

   ENDP  CRC5
      3     5

so make `ENDP & 0x7` and not `ENDP & 0x1f`. This actually could
be a nasty bug, but the code a few lines above makes sure that
`dat` has exactly 11 bits, so actual incorrect mask does not
make any effect.

This patch makes mask according to the spec.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>